+++ src/clojisr/v1/impl/rserve/printing.clj
@@ -10,3 +10,4 @@
                           (prot/eval-r->java session))]
     (->> (.asStrings output)       
          (string/join "\n"))))
+