+++ src/clojisr/v1/r.clj
@@ -1,4 +1,5 @@
 (ns clojisr.v1.r
+  (:refer-clojure :exclude [require])
   (:require [clojisr.v1.protocols :as prot]
             [clojisr.v1.session :as session]
             [clojisr.v1.eval :as evl]
@@ -10,7 +11,7 @@
             [clojisr.v1.impl.clj-to-java :as clj2java]
             [clojure.string :as string]
             [clojisr.v1.help :as help]
-            [clojisr.v1.util :refer [bracket-data maybe-wrap-backtick]]
+            [clojisr.v1.util :refer [maybe-wrap-backtick]]
             [clojisr.v1.require :refer [require-r-package]]
             [clojisr.v1.engines :refer [engines]]
             [clojisr.v1.robject :as robject])
@@ -48,14 +49,19 @@
   (let [session (session/fetch-or-make session-args)]
     (clj2java/clj->java session clj-object)))
 
-(def clj->java->r (comp java->r clj->java))
-(def clj->r clj->java->r)
+(defn clj->java->r [clj-object & {:keys [session-args]}]
+  (let [session (session/fetch-or-make session-args)]
+    (java->r (clj2java/clj->java session clj-object))))
+
+(defn clj->r [clj-object & {:keys [session-args]}]
+  (let [session (session/fetch-or-make session-args)]
+    (java->r (clj2java/clj->java session clj-object))))
 
 (defn r->java->clj [r-object] (-> r-object r r->java java2clj/java->clj))
-(def r->clj r->java->clj)
+(defn r->clj [r-object] (r->java->clj r-object))
 
 (defn r->java->native-clj [r-object] (-> r r-object r->java java2clj/java->native))
-(def r->native-clj r->java->native-clj)
+(defn r->native-clj [r-object] (r->java->native-clj r-object))
 
 (defn session-types
   "Return registered session types"
@@ -76,13 +82,8 @@
 (defn discard-all-sessions []
   (session/discard-all))
 
-(defn apply-function [r-function args & {:keys [session-args]}]
-  (let [session (session/fetch-or-make session-args)]
-    (functions/apply-function r-function args session)))
-
 (defn require-r 
-  "
-   Requires R packages and creates 2 clojure namespaces with functions for each R object per package.
+  "Requires R packages and creates 2 clojure namespaces with functions for each R object per package.
    The 2 namespaces get names of 'package-name' and 'r.package-name'.
 
    It supports as well :as and :refer as clojure.core/require does.
@@ -90,38 +91,36 @@
    so IDEs can show it. 
    
    As this is slow (several seconds for larger packages), it is not enabled by default. It can be enabled
-   using ':generate-doc-strings?' true in the package spec.
+   using ':docstrings?' true in the package spec.
    
    Examples:
 
    (r/require-r '[base])             
    (r/require-r '[stats] :as statistics)
-   (r/require-r '[base :as my-base :generate-doc-strings? true])
-
-
-   "
+   (r/require-r '[base :as my-base :docstrings? true])"
   [& packages]
   (run! require-r-package packages))
 
-(def function functions/function)
+(defn require
+  "Requires R packages and creates 2 clojure namespaces with functions for each R object per package.
+   The 2 namespaces get names of 'package-name' and 'r.package-name'.
 
-(defn println-r-lines
-  "Get a sequence of strings, typically corresponding to lines captured from the standard output of R functions, println them sequentially."
-  [r-lines]
-  (doseq [line r-lines]
-    (println line)))
+   It supports as well :as and :refer as clojure.core/require does.
+   The function can as well attach the R help of  R objects as doc string to the created clojure vars,
+   so IDEs can show it. 
+   
+   As this is slow (several seconds for larger packages), it is not enabled by default. It can be enabled
+   using ':docstrings?' true in the package spec.
+   
+   Examples:
 
-(defn r-lines->md
-  "Get a sequence of strings, typically corresponding to lines captured from the standard output of R functions, format them as markdown."
-  [r-lines]
-  (->> r-lines
-       r->clj
-       (string/join "\n")
-       (format "```\n%s\n```")))
+   (r/require '[base])             
+   (r/require '[stats] :as statistics)
+   (r/require '[base :as my-base :docstrings? true])"
+  [& packages]
+  (run! require-r-package packages))
 
-(defn r-object? [obj]
-  (instance? RObject obj))
-
+(defn r-object? [obj]  (instance? RObject obj))
 
 (comment (defn na [& {:keys [session-args]}]
            (r "NA" :session-args session-args)))
@@ -131,13 +130,13 @@
            empty-symbol
            (r "(quote(f(,)))[[2]]")))
 
-(defn library [libname]
-  (->> libname
-       (format "library(%s)")
-       r))
+(defn library
+  "Load R library in the R runtime."
+  [libname]
+  (->> libname (format "library(%s)") r))
 
 (defn data
-  "Load R dataset and def global var"
+  "Load R dataset and def a global var."
   ([dataset-name] (data dataset-name nil))
   ([dataset-name package]
    (let [n (name dataset-name)
@@ -146,25 +145,19 @@
      (r (format fmt n (name package)))
      (intern *ns* ns (r ns)))))
 
+(defn function?
+  "Checks if given r-object is a function."
+  [r-object]
+  (using-sessions/function? r-object))
 
- (defn- captured-str []
-   "For the R function [str](https://www.rdocumentation.org/packages/utils/versions/3.6.1/topics/str), we capture the standard output and return the corresponding string."
-   (r "function(x) capture.output(str(x))")  )
+(defn object-structure
+  "For the R function [str](https://www.rdocumentation.org/packages/utils/versions/3.6.1/topics/str), we capture the standard output and return the corresponding string."
+  [x]
+  (->> ((r "function(x) capture.output(str(x))") x)
+       (r->clj)
+       (string/join "\n")))
 
- (defn println-captured-str[x] 
-   (->
-    (apply-function
-     (captured-str)
-     [x])
-    println-r-lines))
-
- (defn str-md [x]
-   (->
-    (apply-function
-     (captured-str)
-     [x])
-    r-lines->md))
-
+;;
 
 (defmacro defr
   "Create Clojure and R bindings at the same time"
@@ -175,66 +168,40 @@
      '~name))
 
 (defn rsymbol
-  "Create RObject representing symbol"
+  "Create RObject representing symbol, allows access to private symbols."
   ([string-or-symbol]
    (r (maybe-wrap-backtick string-or-symbol)))
   ([package string-or-symbol]
-   (r (str (maybe-wrap-backtick package) "::" (maybe-wrap-backtick string-or-symbol)))))
-
-
-
-;; FIXME! Waiting for session management.
-(defn- prepare-args-for-bra
-  ([pars]
-   (mapv #(if (nil? %) (r "(quote(f(,)))[[2]]") %) pars))
-  ([pars all?]
-   (if all?
-     (prepare-args-for-bra pars)
-     (conj (prepare-args-for-bra (butlast pars)) (last pars)))))
-
-
-
-
-;; register shutdown hook
-;; should be called once
-(defonce ^:private shutdown-hook-registered
-  (do (.addShutdownHook (Runtime/getRuntime) (Thread. #(locking session/sessions (discard-all-sessions))))
-      true))
-
+   (rsymbol package string-or-symbol false))
+  ([package string-or-symbol private?]
+   (r (str (maybe-wrap-backtick package) (if private? ":::" "::") (maybe-wrap-backtick string-or-symbol)))))
 
 (defn help
   "Gets help for an R object or function"
-  ([r-object]
-   (help/help r-object))
-
-  ([function package]
-   (help/get-help function package)))
-
-
-(defn print-help
-  "Prints help for an R object or function"
-  ([r-object] (println (help r-object)))
-  ([function package] (println (help function package))))
-
+  ([r-object] (help/help r-object))
+  ([function package] (help/get-help function package)))
 
 ;; arithmetic operators
 (defn r- 
   "R arithmetic operator `-`"
-  [e1 e2] ((r "`-`") e1 e2))
+  ([e] ((r "`-`") e))
+  ([e1 e2] ((r "`-`") e1 e2)))
 
 (defn rdiv 
   "R arithmetic operator `/`"
-  [e1 e2] ((r "`/`") e1 e2))
+  ([e] ((r "`/`") 1.0 e))
+  ([e1 e2] ((r "`/`") e1 e2)))
 
 (defn r* 
   "R arithmetic operator `*`, but can be used on an arbitraty number of arguments."
-  [& args] 
-  (reduce (r "`*`") args))
+  [e1 e2 & args] 
+  (reduce (r "`*`") (conj args e2 e1)))
 
 (defn r+
   "R arithmetic operator `+`, but can be used on an arbitraty number of arguments."
-  [& args]
-  (reduce (r "`+`") args))
+  ([e] ((r "`+`") e))
+  ([e & args]
+   (reduce (r "`+`") e args)))
 
 (defn r** 
   "R arithmetic operator `^`"
@@ -306,51 +273,137 @@
 (defn colon 
   "R colon operator `:`"
   [e1 e2] ((r "`:`") e1 e2))
+
 (defn rcolon 
   "R colon operator `:`"
   [e1 e2] (colon e1 e2))
 
+;; tilde
+
+(defn tilde
+  "R ~/formula operator"
+  ([e] (r `(tilde ~e)))
+  ([e1 e2] (r `(tilde ~e1 ~e2))))
+
+(defn rtilde
+  "R ~/formula operator"
+  ([e] (r `(tilde ~e)))
+  ([e1 e2] (r `(tilde ~e1 ~e2))))
+
+;;
+
+(defn r%*%
+  "R matrix product"
+  ([e1] e1)
+  ([e1 e2] ((r "`%*%`") e1 e2)))
+
+(defn r%o%
+  "R outer product"
+  ([e1] e1)
+  ([e1 e2] ((r "`%o%`") e1 e2)))
+
+(defn r%x%
+  "R Kronecker product"
+  ([e1] e1)
+  ([e1 e2] ((r "`%x%`") e1 e2)))
+
 ;; extract/replace operators
+
 (defn r$
   "R extract operator `$`"
   [e1 e2] ((r "`$`") e1 e2))
-
 
 (defn r%in%
   "R match operator `%in%`"
   [e1 e2] ((r "`%in%`") e1 e2))
 
+(defn r%||%
+  "R null coalescing operator `%||%`"
+  [e1 e2] ((r "`%||%`") e1 e2))
 
+;;
+
+(defn- prepare-args-for-bra
+  ([pars]
+   (mapv #(if (nil? %) (r "(quote(f(,)))[[2]]") %) pars))
+  ([pars all?]
+   (if all?
+     (prepare-args-for-bra pars)
+     (conj (prepare-args-for-bra (butlast pars)) (last pars)))))
 
 (defn bra 
   "R extract operator `[`"
   [& pars]
-  (let
-   [bra (clojisr.v1.r/r "`[`")
-    fixed (prepare-args-for-bra pars true)]
-    (clojure.core/apply bra fixed)))
+  (let [bra (r "`[`")
+        fixed (prepare-args-for-bra pars true)]
+    (apply bra fixed)))
 
 (defn brabra 
   "R extract operator `[[`"
   [& pars]
-  (let
-   [bra (clojisr.v1.r/r "`[[`")
-    fixed (prepare-args-for-bra pars true)]
-    (clojure.core/apply bra fixed)))
+  (let [bra (r "`[[`")
+        fixed (prepare-args-for-bra pars true)]
+    (apply bra fixed)))
 
 (defn bra<- 
   "R replace operator `[<-`"
   [& pars]
-  (let
-   [bra (clojisr.v1.r/r "`[<-`")
-    fixed (prepare-args-for-bra pars false)]
-    (clojure.core/apply bra fixed)))
+  (let [bra (r "`[<-`")
+        fixed (prepare-args-for-bra pars false)]
+    (apply bra fixed)))
 
 (defn brabra<- 
   "R replace operator `[[<-`"
   [& pars]
-  (let
-   [bra (clojisr.v1.r/r "`[[<-`")
-    fixed (prepare-args-for-bra pars false)]
-    (clojure.core/apply bra fixed)))
+  (let [bra (r "`[[<-`")
+        fixed (prepare-args-for-bra pars false)]
+    (apply bra fixed)))
 
+;; register shutdown hook
+;; should be called once
+(defonce ^:private _shutdown-hook-registered
+  (do (.addShutdownHook (Runtime/getRuntime) (Thread. #(locking session/sessions (discard-all-sessions))))
+      true))
+
+;; deprecated stuff
+
+(defn println-r-lines
+  "Get a sequence of strings, typically corresponding to lines captured from the standard output of R functions, println them sequentially."
+  {:deprecated true}
+  [r-lines]
+  (doseq [line r-lines]
+    (println line)))
+
+(defn r-lines->md
+  "Get a sequence of strings, typically corresponding to lines captured from the standard output of R functions, format them as markdown."
+  {:deprecated true}
+  [r-lines]
+  (->> r-lines
+       r->clj
+       (string/join "\n")
+       (format "```\n%s\n```")))
+
+(defn println-captured-str
+  {:deprecated "Use `object-structure` and println the result."}
+  [x]
+  (println (object-structure x)))
+
+(defn str-md
+  {:deprecated "Format `object-structure` result with \"```\n%s\n```\")"}
+  [x]
+  (format "```\n%s\n```" (object-structure x)))
+
+(defn print-help
+  "Prints help for an R object or function"
+  {:deprecated "Call `println` on `help` result"}
+  ([r-object] (println (help r-object)))
+  ([function package] (println (help function package))))
+
+;; [ts] I don't see any reason to expose this function here (this function is internal and creates a wrapper around RObject when it represents R function or any callable).
+(def ^{:deprecated true} function functions/function)
+
+(defn apply-function
+  {:deprecated true}
+  [r-function args & {:keys [session-args]}]
+  (let [session (session/fetch-or-make session-args)]
+    (functions/apply-function r-function args session)))